+++ SWFRecomp/include/swf.hpp
@@ -281,7 +281,12 @@ namespace SWFRecomp
 
 		std::unordered_map<u16, float> font_em_square;  // font_id → EM square size
 		std::unordered_map<u16, std::vector<u16>> font_code_tables;  // font_id → code table (index=glyph, value=char code)
-		std::unordered_map<u16, std::vector<s16>> font_advance_tables;  // font_id → per-glyph advance widths
+		// Per-glyph advance widths. The DefineFont2/3 layout advance is an
+		// UNSIGNED 16-bit value (Ruffle `swf/src/read.rs` `Glyph::advance: u16`),
+		// so a 33000-twip advance must survive as 33000, not −32536. Stored as
+		// s32 so the widened range fits and the "no advance" sentinel (−1)
+		// used by the runtime lookups stays out of band.
+		std::unordered_map<u16, std::vector<s32>> font_advance_tables;  // font_id → per-glyph advance widths
 		std::unordered_map<u16, size_t> font_glyph_bases;  // font_id → base index in global glyph_data
 		std::unordered_map<u16, std::string> font_names;  // font_id → font name string
 		std::unordered_map<u16, bool> font_bold_flags;    // font_id → bold flag
